app/requirements.txt
fastapi==0.135.3
uvicorn==0.44.0
pydantic==2.13.0
python-multipart==0.0.20
sqlalchemy==2.0.49
pymysql==1.1.2
numpy==2.3.4
pypng==0.20220715.0
